+++ src/main/java/com/avalon/avalonchat/controller/ProfileController.java
@@ -1,4 +1,4 @@
-package com.avalon.avalonchat.domain.profile.controller;
+package com.avalon.avalonchat.controller;
 
 import static com.avalon.avalonchat.global.util.ResponseEntityUtil.*;
